Filter inactive helm release deletion for only pr releases

+++ scripts/delete-inactive-helm-releases.sh
@@ -44,7 +44,7 @@ for ns in $(echo ${!namespaceMapping[*]}); do
         cutoffDays=${defaultInactiveDays}
       fi
       cutoff=$((cutoffDays*24*3600))
-      if [ $((currenttime-lastUpdated)) -gt "$cutoff" ]
+      if [ $((currenttime-lastUpdated)) -gt "$cutoff" && ${releaseName} = "*-pr-*" ]
        then
          echo "Deleting helm release ${releaseName} as it is inactive for more than ${cutoffDays} days. Last updated : ${date} "
          helm delete --namespace "${ns}" "${releaseName}"
